Update Readme, Language File, rename files, clean up
Follow the main branch
