--- JSON Sample/dict.json
@@ -1,65 +1,0 @@
-[
-  {
-    "image_urls": {
-      "small": "http://i1.pixiv.net/c/150x150/img-master/img/2015/06/24/00/07/22/51055528_p0_master1200.jpg",
-      "large": "http://i1.pixiv.net/img-original/img/2015/06/24/00/07/22/51055528_p0.jpg",
-      "medium": "http://i1.pixiv.net/c/600x600/img-master/img/2015/06/24/00/07/22/51055528_p0_master1200.jpg",
-      "px_480mw": "http://i1.pixiv.net/c/480x960/img-master/img/2015/06/24/00/07/22/51055528_p0_master1200.jpg",
-      "px_128x128": "http://i1.pixiv.net/c/128x128/img-master/img/2015/06/24/00/07/22/51055528_p0_square1200.jpg"
-    },
-    "type": "illustration",
-    "favorite_id": 0,
-    "width": 1000,
-    "stats": {
-      "favorited_count": {
-        "private": 469,
-        "public": 5767
-      },
-      "views_count": 139847,
-      "commented_count": 175,
-      "score": 53870,
-      "scored_count": 5466
-    },
-    "book_style": "right_to_left",
-    "content_type": null,
-    "is_liked": false,
-    "metadata": null,
-    "age_limit": "all-age",
-    "user": {
-      "is_following": true,
-      "name": "KD",
-      "id": 395595,
-      "account": "cadillac",
-      "is_friend": false,
-      "profile": null,
-      "is_follower": false,
-      "profile_image_urls": {
-        "px_50x50": "http://i2.pixiv.net/img20/profile/cadillac/8457709_s.jpg"
-      },
-      "is_premium": null,
-      "stats": null
-    },
-    "title": "ゆき",
-    "sanity_level": "white",
-    "reuploaded_time": "2015-06-24 00:07:22",
-    "tools": [
-      "Photoshop",
-      "SAI"
-    ],
-    "id": 51055528,
-    "publicity": 0,
-    "created_time": "2015-06-24 00:07:22",
-    "page_count": 1,
-    "is_manga": false,
-    "height": 1414,
-    "caption": "感谢Facebook网友的写真参考https://www.facebook.com/profile.php?id=1272330679\r\nWorldCosplay- http://worldcosplay.net/photo/3499488",
-    "tags": [
-      "女子高生",
-      "オリジナル",
-      "セーラー服",
-      "黒髪ロング",
-      "模写",
-      "オリジナル5000users入り"
-    ]
-  }
-]
