--- JSON Sample/illustrations_data.json
@@ -1,142 +1,0 @@
-[
-  {
-    "image_urls": {
-      "small": "http://i4.pixiv.net/c/150x150/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-      "large": "http://i3.pixiv.net/img-original/img/2015/07/09/00/05/01/51320366_p0.jpg",
-      "medium": "http://i3.pixiv.net/c/600x600/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-      "px_480mw": "http://i3.pixiv.net/c/480x960/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-      "px_128x128": "http://i3.pixiv.net/c/128x128/img-master/img/2015/07/09/00/05/01/51320366_p0_square1200.jpg"
-    },
-    "type": "illustration",
-    "favorite_id": 0,
-    "width": 800,
-    "stats": {
-      "favorited_count": {
-        "private": 236,
-        "public": 3246
-      },
-      "views_count": 62043,
-      "commented_count": 54,
-      "score": 25741,
-      "scored_count": 2600
-    },
-    "book_style": "none",
-    "content_type": "none",
-    "is_liked": false,
-    "metadata": "none",
-    "age_limit": "all-age",
-    "user": {
-      "is_following": true,
-      "name": "KD",
-      "id": 395595,
-      "account": "cadillac",
-      "is_friend": false,
-      "profile": null,
-      "is_follower": false,
-      "profile_image_urls": {
-        "px_50x50": "http://i2.pixiv.net/img20/profile/cadillac/8457709_s.jpg"
-      },
-      "is_premium": null,
-      "stats": null
-    },
-    "title": "ドルフィンシンフォニー",
-    "sanity_level": "white",
-    "reuploaded_time": "2015-07-09 00:05:01",
-    "tools": [
-      "Photoshop",
-      "SAI"
-    ],
-    "id": 51320366,
-    "publicity": 0,
-    "created_time": "2015-07-09 00:05:01",
-    "page_count": 1,
-    "is_manga": false,
-    "height": 1131,
-    "caption": "今年もミニスカとthe太ももの季節がやってきました！",
-    "tags": [
-      "女子高生",
-      "オリジナル",
-      "セーラー服",
-      "イルカ",
-      "黒ハイソックス",
-      "MDR-Z1000",
-      "青",
-      "ヘッドホン"
-    ]
-  },
-  {
-    "mode": "daily",
-    "date": "2015-07-10",
-    "content": "all",
-    "works": [
-      {
-        "rank": 1,
-        "previous_rank": 7,
-        "work": {
-          "image_urls": {
-            "small": "http://i3.pixiv.net/c/150x150/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-            "large": "http://i3.pixiv.net/img-original/img/2015/07/09/00/05/01/51320366_p0.jpg",
-            "medium": "http://i3.pixiv.net/c/600x600/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-            "px_480mw": "http://i3.pixiv.net/c/480x960/img-master/img/2015/07/09/00/05/01/51320366_p0_master1200.jpg",
-            "px_128x128": "http://i3.pixiv.net/c/128x128/img-master/img/2015/07/09/00/05/01/51320366_p0_square1200.jpg"
-          },
-          "type": "illustration",
-          "favorite_id": null,
-          "width": 800,
-          "stats": {
-            "favorited_count": {
-              "private": null,
-              "public": null
-            },
-            "views_count": 42718,
-            "commented_count": null,
-            "score": 6392,
-            "scored_count": 644
-          },
-          "book_style": "none",
-          "content_type": null,
-          "is_liked": null,
-          "metadata": null,
-          "age_limit": "all-age",
-          "user": {
-            "is_following": null,
-            "name": "KD",
-            "id": 395595,
-            "account": "cadillac",
-            "is_friend": null,
-            "profile": null,
-            "is_follower": null,
-            "profile_image_urls": {
-              "px_170x170": "http://i2.pixiv.net/img20/profile/cadillac/8457709.jpg",
-              "px_50x50": "http://i2.pixiv.net/img20/profile/cadillac/8457709_s.jpg"
-            },
-            "is_premium": null,
-            "stats": null
-          },
-          "title": "ドルフィンシンフォニー",
-          "sanity_level": "white",
-          "reuploaded_time": "2015-07-09 00:05:01",
-          "tools": null,
-          "id": 51320366,
-          "publicity": 0,
-          "created_time": "2015-07-09 00:05:00",
-          "page_count": 1,
-          "is_manga": null,
-          "height": 1131,
-          "caption": null,
-          "tags": [
-            "女子高生",
-            "オリジナル",
-            "セーラー服",
-            "イルカ",
-            "黒ハイソックス",
-            "MDR-Z1000",
-            "青",
-            "ヘッドホン"
-          ]
-        }
-      },
-      "XXX"
-    ]
-  }
-]
