+++ locale/zh_CN/LC_MESSAGES/messages.po
@@ -1,0 +1,142 @@
+# SOME DESCRIPTIVE TITLE.
+# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
+# This file is distributed under the same license as the PACKAGE package.
+# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
+#
+#, fuzzy
+msgid ""
+msgstr ""
+"Project-Id-Version: PACKAGE VERSION\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-01-08 12:06+0800\n"
+"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
+"Language-Team: LANGUAGE <LL@li.org>\n"
+"Language: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+
+#: Pixiv Downloader.py:138
+#, python-format
+msgid "Connection error: %s"
+msgstr "连接错误: %s"
+
+#: Pixiv Downloader.py:140
+msgid "process cancelled"
+msgstr "过程已中断"
+
+#: Pixiv Downloader.py:142
+#, python-format
+msgid "%s => %s download failed"
+msgstr "%s => %s 下载失败"
+
+#: Pixiv Downloader.py:144
+#, python-format
+msgid "%s => %s download error, retry"
+msgstr "%s => %s 下载错误，重试"
+
+#: Pixiv Downloader.py:200
+msgid "Start download, total illustrations "
+msgstr "开始下载，插画总数 "
+
+#: Pixiv Downloader.py:216
+msgid "There is no new illustration need to download"
+msgstr "没有新的插画需要下载"
+
+#: Pixiv Downloader.py:221
+msgid "Input the artist's id:(separate with space)"
+msgstr "输入艺术家的id：(用空格分隔)"
+
+#: Pixiv Downloader.py:223 Pixiv Downloader.py:264
+#, python-format
+msgid "Artists %s ??\n"
+msgstr "艺术家 %s ??\n"
+
+#: Pixiv Downloader.py:236
+msgid "Input the date:(eg:2015-07-10)"
+msgstr "输入日期：(例:2015-07-10)"
+
+#: Pixiv Downloader.py:238
+msgid "[invalid]"
+msgstr "[失效]"
+
+#: Pixiv Downloader.py:262
+#, python-format
+msgid "Artists %s\n"
+msgstr "艺术家 %s\n"
+
+#: Pixiv Downloader.py:273
+msgid " Pixiv Downloader 2.2 "
+msgstr " Pixiv 下载器 2.2 "
+
+
+#: Pixiv Downloader.py:284
+msgid "Which do you want to:"
+msgstr "功能选择："
+
+#: Pixiv Downloader.py:287
+msgid "exit"
+msgstr "退出"
+
+#: Pixiv Downloader.py:289
+msgid "download_by_user_id"
+msgstr "下载指定艺术家作品"
+
+#: Pixiv Downloader.py:290
+msgid "download_by_ranking"
+msgstr "下载今日排行榜"
+
+#: Pixiv Downloader.py:291
+msgid "download_by_history_ranking"
+msgstr "下载历史排行榜"
+
+#: Pixiv Downloader.py:292
+msgid "update_exist"
+msgstr "更新已存在艺术家作品"
+
+#: Pixiv Downloader.py:293
+msgid "refresh_exist"
+msgstr "刷新已存在艺术家作品"
+
+#: Pixiv Downloader.py:295
+msgid " finished "
+msgstr " 完成"
+
+#: Pixiv Downloader.py:299
+msgid "Wrong input!"
+msgstr "输入错误"
+
+#: api.py:49
+msgid "Checking session"
+msgstr "验证会话"
+
+#: api.py:63
+msgid " [VALID]"
+msgstr " [有效]"
+
+#: api.py:65
+msgid " [EXPIRED]"
+msgstr " [无效]"
+
+#: api.py:106
+msgid "Unknown Method:"
+msgstr "未知方法:"
+
+#: api.py:147
+msgid "[ERROR] connection failed!"
+msgstr "[错误] 连接失败!"
+
+#: api.py:151
+msgid "Please login"
+msgstr "请登录"
+
+#: api.py:152
+msgid "username:"
+msgstr "用户名"
+
+#: api.py:153
+msgid "password:"
+msgstr "密码"
+
+
